Process RoundStateResponse according to backend version
